requirements.txt
 flask 
 flask-sqlalchemy 
 flask-cors 
 boto3 
 python-dotenv 
 requests